fix(tests): Change parseInt to parseFloat for solAmount conversion
**Issue:** TypeScript compilation errors in tests 04-09

**Root Cause:** parseInt() returns number but PRODUCTION_CONFIG.swapAmount is a string like '0.01', so parseInt('0.01') returns 0, not 10000000. Also, the multiplication result needs to match the expected type.

**Fix:** Use parseFloat() instead of parseInt() to properly convert decimal SOL values to lamports.

**Files Changed:**
- tests/production/e2e/04-agreement-expiry-refund.test.ts
- tests/production/e2e/05-admin-cancellation.test.ts
- tests/production/e2e/06-zero-fee-transactions.test.ts
- tests/production/e2e/07-idempotency-handling.test.ts
- tests/production/e2e/09-edge-cases-validation.test.ts (2 occurrences)

**Example:**
- Before: parseInt('0.01') * 1_000_000_000 = 0 * 1_000_000_000 = 0 ❌
- After: parseFloat('0.01') * 1_000_000_000 = 0.01 * 1_000_000_000 = 10_000_000 ✅

+++ tests/production/e2e/04-agreement-expiry-refund.test.ts
@@ -119,7 +119,7 @@ describe('PRODUCTION E2E: Agreement Expiry and Refund', function () {
     const requestBody = {
       nftMint: expiryNft.mint.toString(),
       swapType: 'NFT_FOR_SOL',
-      solAmount: parseInt(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
+      solAmount: parseFloat(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
       seller: wallets.sender.publicKey.toString(),
       buyer: wallets.receiver.publicKey.toString(),
       expiry: expiry.toISOString(),
