+++ tests/production/e2e/09-edge-cases-validation.test.ts
@@ -97,7 +97,7 @@ describe('PRODUCTION E2E: Edge Cases and Validation', function () {
         {
           nftMint: invalidMint.toString(),
           swapType: 'NFT_FOR_SOL',
-          solAmount: parseInt(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
+          solAmount: parseFloat(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
           seller: wallets.sender.publicKey.toString(),
           buyer: wallets.receiver.publicKey.toString(),
           expiry: expiry.toISOString(),
@@ -226,7 +226,7 @@ describe('PRODUCTION E2E: Edge Cases and Validation', function () {
         {
           nftMint: testNft.mint.toString(),
           swapType: 'NFT_FOR_SOL',
-          solAmount: parseInt(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
+          solAmount: parseFloat(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
           seller: wallets.sender.publicKey.toString(),
           buyer: wallets.receiver.publicKey.toString(),
           expiry: expiry.toISOString(),
